+++ src/modules/Product/types/Presentation.ts
@@ -1,0 +1,11 @@
+export interface Presentation {
+    sale_type: number;
+    int_cod: string;
+    bar_cod: string;
+    packing_deployed: string;
+    packing_json: string;
+    stock_min: number;
+    stock_max: number;
+    price: string;
+    status: number;
+  }
