+++ src/types/RadioOption.ts
@@ -1,0 +1,4 @@
+export interface RadioOption {
+    label: string;
+    value: number;
+}
